src/nlinkfs.c: Do not show duplicated files
Given an abcd.LNK file in the source directory, if an abcd file exists,
nlinkfs will show (wrongly) two link entries in the directory. This commit
fixes this behavior and takes the following approach:

If two files: file and file.LNK exist in the same directory, nlinkfs will
consider the file.LNK as a regular file and will not create the link, even
if the file.LNK it's a valid nlinkfs link file.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>